backend/requirements.txt
Flask~=3.0.0
Flask-SQLAlchemy~=3.1.1
Flask-JWT-Extended~=4.5.3
Flask-CORS~=4.0.0
Flask-Limiter~=3.5.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0
werkzeug~=3.0.1
GitPython~=3.1.40
pytest~=7.4.3
pytest-flask~=1.3.0
gunicorn~=21.2.0
anthropic~=0.40.0
